feat: Process original email to display formatted emails
