+++ .vscode/cspell.json
@@ -5,6 +5,7 @@
   "dictionaries": [],
   "words": [
     "Authed",
+    "bullmq",
     "Commenters",
     "composables",
     "convo",
