+++ apps/mail-bridge/utils/purify.ts
@@ -1,0 +1,8 @@
+import DOMPurify from 'dompurify';
+import { JSDOM } from 'jsdom';
+
+const window = new JSDOM('').window;
+const purify = DOMPurify(window);
+
+export const sanitize = (html: string) =>
+  purify.sanitize(html, { USE_PROFILES: { html: true } });
