+++ packages/otel/exports.ts
@@ -1,0 +1,1 @@
+export { trace } from '@opentelemetry/api';
